+++ test/test_helper.exs
@@ -8,6 +8,8 @@ Mimic.copy(Trenino.Simulator)
 Mimic.copy(Trenino.Simulator.Connection)
 Mimic.copy(Trenino.Simulator.AutoConfig)
 Mimic.copy(Trenino.AppVersion)
+Mimic.copy(Circuits.UART, type_check: true)
+Mimic.copy(Trenino.Serial.Discovery, type_check: true)
 
 ExUnit.start()
 Ecto.Adapters.SQL.Sandbox.mode(Trenino.Repo, :manual)